fix: cast float BITS to usize for float_bits_const compat
When `float_bits_const` stabilizes, `f32::BITS` / `f64::BITS` will be
inherent consts returning `u32`, shadowing the trait's `AsBytes::BITS`
(which returns `usize`). Add `as usize` to the atomic float `AsBytes`
impl in `impl_float!` to prevent a type mismatch.

Fixes #7

+++ src/impls.rs
@@ -1008,7 +1008,7 @@ macro_rules! impl_float {
     ($($ty:ty, $aty:ty, $zero:expr, $one:expr,)*) => {$(
 
 impl AsBytes for $aty {
-    const BITS: usize = <$ty>::BITS;
+    const BITS: usize = <$ty>::BITS as usize;
     const BYTES: usize = <$ty>::BYTES;
     type Bytes = [u8;  <$ty>::BYTES];
 }
